Support multi-provider SSO profile linking

+++ includes/class-user-manager.php
@@ -59,8 +59,9 @@ class AOAUTH_User_Manager {
             }
 
             $linked_provider = get_user_meta($existing_user->ID, '_aoauth_provider', true);
+            $linked_providers = AOAUTH_Core::get_user_linked_providers($existing_user->ID);
             
-            if ($linked_provider === $provider_slug) {
+            if ($linked_provider === $provider_slug || isset($linked_providers[$provider_slug])) {
                 return $this->login_existing_user($existing_user, $user_info, $provider, $provider_slug);
             }
             
@@ -307,8 +308,14 @@ class AOAUTH_User_Manager {
             $security->clear_linking_attempts($user->ID, $linking_key);
         }
         
-        update_user_meta($user->ID, '_aoauth_provider', $linking_data['provider_slug']);
-        update_user_meta($user->ID, '_aoauth_linked_' . $linking_data['provider_slug'], time());
+        $provider_user_data = isset($linking_data['user_data']) && is_array($linking_data['user_data']) ? $linking_data['user_data'] : array();
+        $provider_subject = $this->get_provider_subject($provider_user_data);
+        $provider_email = isset($provider_user_data['email']) ? sanitize_email($provider_user_data['email']) : '';
+        if (AOAUTH_Core::provider_identity_belongs_to_other_user($user->ID, $linking_data['provider_slug'], $provider_email, $provider_subject)) {
+            return new WP_Error('account_linking_identity_in_use', __('This provider account is already linked to another WordPress user.', 'aoauth-client-sso'));
+        }
+
+        AOAUTH_Core::link_user_provider($user->ID, $linking_data['provider_slug'], $provider_email, $provider_subject);
         
         delete_transient($linking_key);
         
@@ -327,10 +334,8 @@ class AOAUTH_User_Manager {
             return new WP_Error('login_blocked', 'Login blocked by administrator.');
         }
         
-        $current_provider = get_user_meta($user->ID, '_aoauth_provider', true);
-        if (empty($current_provider)) {
-            update_user_meta($user->ID, '_aoauth_provider', $provider_slug);
-        }
+        $provider_subject = $this->get_provider_subject($user_info);
+        AOAUTH_Core::link_user_provider($user->ID, $provider_slug, $user_info['email'] ?? '', $provider_subject);
         
         update_user_meta($user->ID, '_aoauth_last_login', time());
         
@@ -400,12 +405,22 @@ class AOAUTH_User_Manager {
             return new WP_Error('user_creation_failed', 'Failed to create user account. Please try again.');
         }
         
-        update_user_meta($user_id, '_aoauth_provider', $provider_slug);
+        AOAUTH_Core::link_user_provider($user_id, $provider_slug, $user_info['email'] ?? '', $this->get_provider_subject($user_info));
         update_user_meta($user_id, '_aoauth_created', time());
         
         do_action('aoauth_user_created', $user_id, $user_info, $provider);
         
         return $user_id;
+    }
+
+    private function get_provider_subject($user_info) {
+        foreach (array('subject', 'sub', 'id', 'user_id') as $subject_key) {
+            if (!empty($user_info[$subject_key])) {
+                return sanitize_text_field((string) $user_info[$subject_key]);
+            }
+        }
+
+        return '';
     }
     
     private function generate_unique_username($username) {
